파일 이름 수정, useMediaCheck.tsx 수정
