+++ src/components/Market/ProductDetails.tsx
